requirements.txt
asgiref==3.7.2
Django==4.2.7
django-braces==1.15.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.6.0
